Transfered trigger to trigger subsystem (from shooter)
