+++ src/main/java/frc/robot/commands/LoadTrigger.java
@@ -8,9 +8,11 @@ package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.subsystems.Shooter;
+import frc.robot.subsystems.Trigger;
 
 public class LoadTrigger extends Command {
   private Shooter shooter;
+  private Trigger trigger;
   /** Creates a new LoadTrigger. */
   public LoadTrigger(Shooter shooter) {
     // Use addRequirements() here to declare subsystem dependencies.
@@ -21,8 +23,8 @@ public class LoadTrigger extends Command {
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
-    shooter.setTopTriggerMotorVelocity(5);
-    shooter.setBottomTriggerMotorVelocity(5);
+    trigger.setTopTriggerMotorVelocity(5);
+    trigger.setBottomTriggerMotorVelocity(5);
   }
 
   // Called every time the scheduler runs while the command is scheduled.
@@ -32,13 +34,13 @@ public class LoadTrigger extends Command {
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    shooter.setTopTriggerMotorVelocity(0);
-    shooter.setBottomTriggerMotorVelocity(0);
+    trigger.setTopTriggerMotorVelocity(0);
+    trigger.setBottomTriggerMotorVelocity(0);
   }
 
   // Returns true when the command should end.
   @Override
   public boolean isFinished() {
-    return shooter.noteIsInTrigger();
+    return trigger.noteIsInTrigger();
   }
 }
