+++ src/main/java/frc/robot/commands/RunShooter.java
@@ -8,9 +8,11 @@ package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.subsystems.Shooter;
+import frc.robot.subsystems.Trigger;
 
 public class RunShooter extends Command {
   private Shooter shooter;
+  private Trigger trigger;
   private double shooterTopRPS;
   private double shooterBottomRPS;
   private double triggerTopRPS;
@@ -41,8 +43,8 @@ public class RunShooter extends Command {
     /* if shooter motors are up to speed, then turn on trigger motors */
     /* might need to adjust the numbers depending on what % we want the power at */
     if ((shooter.getTopShooterMotorVelocity() >= 0.98 * shooterTopRPS) && (shooter.getBottomShooterMotorVelocity() >= 0.98 * shooterBottomRPS)){
-      shooter.setTopTriggerMotorVelocity(triggerTopRPS);
-      shooter.setBottomShooterMotorVelocity(triggerBottomRPS);
+      trigger.setTopTriggerMotorVelocity(triggerTopRPS);
+      trigger.setBottomTriggerMotorVelocity(triggerBottomRPS);
     }
   }
 
@@ -51,14 +53,14 @@ public class RunShooter extends Command {
   public void end(boolean interrupted) {
     shooter.setTopShooterMotorVelocity(0);
     shooter.setBottomShooterMotorVelocity(0);
-    shooter.setTopTriggerMotorVelocity(0);
-    shooter.setBottomTriggerMotorVelocity(0);    
+    trigger.setTopTriggerMotorVelocity(0);
+    trigger.setBottomTriggerMotorVelocity(0);    
   }
 
   // Returns true when the command should end.
   /* return true when the note is not in the trigger/shooter */
   @Override
   public boolean isFinished() {
-    return !shooter.noteIsInShooter() && !shooter.noteIsInTrigger();
+    return !shooter.noteIsInShooter() && !trigger.noteIsInTrigger();
   }
 }
