+++ src/main/java/frc/robot/subsystems/Pivot.java
@@ -8,7 +8,6 @@ import com.ctre.phoenix6.configs.TalonFXConfiguration;
 import com.ctre.phoenix6.controls.PositionVoltage;
 import com.ctre.phoenix6.hardware.TalonFX;
 import edu.wpi.first.util.sendable.SendableBuilder;
-
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 
 public class Pivot extends Diagnostics {
